Change some data type

+++ VaccineScheduleTracking.API/Models/Entities/Child.cs
@@ -5,8 +5,8 @@
         public int ChildID { get; set; }
         public string Firstname { get; set; }
         public string Lastname { get; set; }
-        public double Weight { get; set; }
-        public double Height { get; set; }
+        public decimal Weight { get; set; }
+        public decimal Height { get; set; }
         public string Gender { get; set; }
         public DateOnly DateOfBirth { get; set; }
         public int ParentID { get; set; }
